Move denoising to dedicated package.
